reactor: simplify and clarify pollers
As we have found out with IO experiments in the past, the order
of poller execution is important for performance. It is not hard
to see why: one poller may generate events that others want to
consume. If the consumption doesn't happen right away we have to
wait until the next task quota.

The way things are right now, we don't control when the epoll
poller run. It also means that the backend has to do all the work
of dealing with that poller.

This seems to happen because we apply an optimization that only
runs the epoll poller when epoll is used. However, I will argue
that this is not worth the hassle. That is especially true given
that the optimization doesn't ever stop polling for epoll once
it is used once. In practice we can expect the epoll poller to
always be used once in any complex software.

When doing that we also decouple the role of the wait_and_process
method. Right now it is used both for epoll polling and going to
sleep. But this gets confusing once the epoll poller is now managed
by the reactor. On top of that, io_uring, which is coming soon, has
no need for wait_and_process for polling events: just to go to sleep.

Another advantage of this approach is that it will simplify the
backend code and give the backends more options on how to handle
their own completions without having to deal with register pollers.

Signed-off-by: Glauber Costa <glauber@scylladb.com>