liburing: add seastar-specific class to deal with liburing
Liburing is a library that implements functionality to deal with the new
Linux asynchronous invocation API (uring)

Before implementing support for it, this patch will add a seastar class,
uring_context, that implements the functionality that we need.

As far as cmake goes, liburing will be completely optional (note that
in CMakeLists, we don't have a fatal statement in case it is not found).
If it is found, it will be used. If not found, not used.

Signed-off-by: Glauber Costa <glauber@scylladb.com>