reactor: rework aio completions and factor out preempt io context
The uring implementation will still use aio to implement the task quota.
In order to achieve that, it would be nice to factor out that code so it
can be reused by uring. This patch implements such work.

The first step is to move the aio_context out of reactor_backend_aio.
It is now called aio_general_context, as we already have (for now) one
I/O-specific context.

That context can now be used in another class, preempt_io_context, that
implements the task quota.

Two complications arise when doing that, both steming from that the aio
code was used by the preempt context and also the polling context:

 1 . One of the iocbs, for the hrtimer, was shared between those two
     contexts with a flag dance so we can tell its current status.
 2 . A bit as consequence of 1, it is hard to change the the code to
     handle iocbs just for the preempt case, since a lot of it is
     shared in the aio_context (like replenish, queue, etc)

To make disentangle this, we will take advantage of the fact that we
can now use kernel_completions to store an action and greatly simplify
the io_getevents code by always relying on kernel_completions and
implementing the relevant actions. Some more details about this process:

1. For the hrtimer, we will have two separate actions: one inside the
   the preempt_io_context, and another specific for the polling context.
   Having an extra iocb helps keep things independent and it's not a big
   waste.

2. For all of the event processing code we will use event specific
   completions like smp_wakeup_aio_completion. Note the _aio_ in the
   name: it's already in anticipation of the fact that uring will
   have its own. The _aio_ versions to the completions have iocbs
   attached directly to it, so the uring code won't have to worry about
   it.

3. For IOCB_POLL, the current code allocates an iocb that it would
   later have to free. io_uring won't have this limitation, and since
   the kernel_completion for poll is now stored inside
   pollable_fd_state, it is better to use the same type for both
   io_uring and aio. Therefore it is better not to allocate. To do that,
   we will move the iocb itself inside the poll's kernel completion.
   Although that sounds wasteful and ugly since it is not generic
   enough, it's better than the alternative.

Signed-off-by: Glauber Costa <glauber@scylladb.com>