reactor: move tcp socket functions to reactor_backend
accept, read_some (scalar and vectored), and write_some (scalar and
vectored) are moved to the reactor backend. The implementation will
be the same for aio and epoll, and as such they are implemented as
a set of common helpers.

A class is used to glue the helpers together just so that we can
make friendship easier in the pollable_fd_state: the helpers are
fully static and don't depend on any state.

Signed-off-by: Glauber Costa <glauber@scylladb.com>