reactor: rework submit_io so that it is backend-agnostic
This patch reworks submit_io so that it only queues an I/O request,
nothing more. Later, at the submit stage, the reactor backend will pick
the best way to submit this. Note please that at this point this still
all live inside the reactor itself.

To do that, I am attaching the kernel_completion pointer inside the
io_request. The io_request is now queued, and has all the information it
needs to be submitted.

Akin to what we are doing in the io_uring patches, we will also not
submit all available I/O during submission: we will only submit up to
the point in which we have available iocbs. The theory behind it is that
if we don't have iocbs available that's because we exhausted our I/O
capacity anyway, and it is pointless to submit.

We can always submit more in the next task quota, when capacity frees
up. Also note that since the previous patch we will always reap events
before we submit, so if something had just completed it would have freed
us slots.

Doing that allow us to remove the semaphore, which is a pre-condition to
moving all of this code inside the pollers.

Signed-off-by: Glauber Costa <glauber@scylladb.com>