reactor: move tcp socket functions to reactor_backend
accept, read_some (scalar and vectored), and write_some (scalar and
vectored) are moved to the reactor backend. The implementation will
be the same for aio and epoll, and as such they are implemented as
a set of common helpers.

There are no users in tree except for a test, that is patched, of
the direct calls to engine().read/write_some(), so to avoid one
extra indirection those functions are just replaced with their
do_ variants and the pollable_fd_state calls the backend directly

Signed-off-by: Glauber Costa <glauber@scylladb.com>