reactor: simplify and clarify pollers
As we have found out with IO experiments in the past, the order
of poller execution is important for performance. It is not hard
to see why: one poller may generate events that others want to
consume. If the consumption doesn't happen right away we have to
wait until the next task quota.

The way things are right now, we don't control when the epoll
poller run. It also means that the backend has to do all the work
of dealing with that poller. Also as we move towards more backends,
like io_uring, we want clearer roles for those pollers. Even
for our current AIO backend, it is possible in principle to use
the same context for storage and poll, and merging those pollers
into one will make this work easier.

So the current situation has:
 - an epoll poller, that runs only when poll() was ever called (
   never, for dpdk, pretty much all the time for posix-network),
   but we never know when
 - an storage I/O kernel completion poller
 - a joint I/O queue and storage I/O kernel submitter.

For storage, we want to keep kernel completions and kernel submissions
separate. This is because completing work may free up slots in the I/O
queue, that in turn may submit the new work right away. This order is
present now, except the I/O poller does both AIO and I/O queue work.

For poll we want submissions to run first and reap later, because some
of the work submitted may be ready right away. This is not guaranteed
now because we don't control when the epoll poller is created.

This patch then splits this into the following pollers:

- 1. kernel completions: reap events that are ready from the kernel.
     Events can be of any kind, storage, poll, etc.
- 2. I/O queue: run the I/O queue.
- 3. kernel submission: submit new work to the kernel.

We want the order of polling to be 1 - 2 - 3 - 1. Since reaping
completions is generally cheap, that shouldn't be a problem.

When doing that we also decouple the role of the wait_and_process
method. Right now it is used both for epoll polling and going to
sleep. But this gets confusing once we merge kernel submission into
its own poller.

Another advantage of this approach is that it will simplify the
backend code and give the backends more options on how to handle
their own completions without having to deal with register pollers.

Signed-off-by: Glauber Costa <glauber@scylladb.com>