io_request: enhance it with more operation types
io_uring has a lot more operation types, and we would like to keep them
all in the same descriptor, so they can be used by the same queue.

Add the operation types, and to prevent this from becoming a type-mess,
use std::variant to store different types, while providing accessors for
all the types.

The prepare_iocb code can already be simplified by doing this, and we
seize this opportunity. A logical next step is to store io_requests
instead of iocb in the aio_general_context queues, which will be a
pre-requisite for joining poll and storage I/O in a single context.
This is deferred for a later time.

Signed-off-by: Glauber Costa <glauber@scylladb.com>