io_request: allow the request to be marked as force-async
Some backends, like io_uring, have the ability to force asynchronous
dispatch. If that's the case, the operation will never be tried inline
within the system call but always dispatched to a thread.

In general it is preferred to do operations inline: some are very
lightweight and there is no point in adding the overhead.  In an ideal
world, we'd try all operations inline.

But our experience with XFS tell us that there are many situations in
which it may block. While the ones we know are fixed, there are likely
others we don't know about.

I am proposing this to be a property of the request, not the operation.
This is because the same operation, like read, can be used in many
contexts, like sockets and file I/O. The I/O queue when dispatching
will mark the request as force-async.

Signed-off-by: Glauber Costa <glauber@scylladb.com>