net: remove smart pointer wrappers around pollable_fd
pollable_fd is now equivalent to shared_ptr, so there is no need to
wrap it with std::unique_ptr or lw_shared_ptr.

A couple of missing methods were added: a default constructor and
bool conversion operator, since they were expected by users.

This reduces allocations and indirections. None of it is critical
since it happens once per socket, but it's still nice.