reactor: generalize common parts of shutdown and close
shutdown was moved to the backend so that some backends, like uring,
could cancel ongoing connect / accept before shutting a socket down.

Some other operations, like oingoing polls, also need to be canceled but
they would be cancelled within forget().

With the new refcount mechanism introduced by Avi, we may never reach
forget(), which is now called when the refcount reaches zero.

The reason is, the operations that are in-flight will themselves hold
refcounts and have to be explicitly cancelled.

To aid in that, we will rename the shutdown() backend function to
cancel_ongoing_operations. It is now called both for shutdown() and
close(), and the actual actions of shutdown and close (calling shutdown
and resetting the pollable_fd_state) are called from common code.

Signed-off-by: Glauber Costa <glauber@scylladb.com>