fair_queue: extract potential common elements to a common class
I am about to introduce a new kind of fair queue to allow nesting.
While we could allow for nesting using individual fair queues, it
is more efficient that nested fair queues be different than leaf
fair queues.

Some parts of it, though, will be common. Parts of the fair queue
and its priority class that I expect to be the same between both
fair queues are now extracted into a common class.

As we move this to common code, some small changes are also carried
out:
 - can_dispatch() changes, because whether or not a queue has requests
   pending will be a leaf queue's detail. We use the fact that there are
   priority classes registered to make this decision.

 - can_dispatch() changes further. As capacity changes will be allowed
   soon, we will rely on new fields for current limits and not on whatever
   is in the _config field.

 - We will use pointers and not shared pointers in the _handlers method,
   so users of methods that deal with it are changed to match.

Signed-off-by: Glauber Costa <glauber@scylladb.com>