fair_queue: allow the fair queues to modify their capacity
This patch introduces a new type, the fair_queue_credit, that will be
used to carry information about capacity changes we can apply to a fair
queue.

Operations are added to add and remove capacity, and also a helper to
remove the entire capacity at once.

Note that these operations are intendended to be public. Where I am
going with this, only nested queues will be able to influence each
other's capacity.

A test is added, and specifically for the test those methods are
made public.

Signed-off-by: Glauber Costa <glauber@scylladb.com>