fair_queue: also keep track of queued weight and size
The fair queue keeps track of how many queued requests we have,
but not their properties (size, weight factor).

This has been okay because we only use that so far to keep a metric
of queued requests. However we will soon use the information of what
is queued to inform a nested queue (upcoming).

We will need full information in that case. We also need to guarantee
that requests do not have zeroed sizes or weights. The reason for that
is that as we try to add or remove capacity based on our queued requests,
if one of our dimensions is zero we will not add anything. That can block
new requests from passing through even if the other dimensions are free.

Signed-off-by: Glauber Costa <glauber@scylladb.com>