fair_queue: bulk_fair_queue, a way to nest fair queues
This patch introduces the bulk_fair_queue, a class similar to the
fair_queue that supports bulk dispatching and is safe to use across
shards.

The motivation for implementing this structure is to allow fair_queues
to be nested and to allow multi-shard dispatch.

In the fair_queue, each request is put in a queue inside its priority
class and that is later dispatched and accounted for. Although it is
far from simple, we can in theory make this queue safe for smp access.

However, the cost of processing many small requests from various sources
quickly shows this approach to be infeasible.

Instead of that, we can allow requests to accumulate in a fair_queue:
many requests, from many different priority_classes. When the fair_queue
wants to dispatch it tells its bulk_fair_queue about its needs. The
bulk_fair_queue then loans capacity back to the fair_queue, which
can then dispatch its own requests.

In the multi-shard case the expectation is that there will be a single
bulk_fair_queue, and each shard will register its own local fair_queue
against it. Each shard is allowed to queue() at any time. Queueing
happens through an spsc queue.

All shards can then call dispatch_requests(). If we think for instance
about the main user of the fair_queue, the I/O scheduler, we have the
following situation:

- each shard accumulates all its local I/O and asks the bulk_fair_queue
  for capacity
- As shards gets to their pollers, they grant the loans for themselves
  and others.
- If many shards try to dispatch at the same time, one of them will
  dispatch and the others will just go on with their lives.

Signed-off-by: Glauber Costa <glauber@scylladb.com>