fair_queue: express all internal fair_queue quantities as fair_queue_tickets
The fair_queue_request have recently evolved to become a
fair_queue_ticket.  But the fair_queue itself still tracks requests
individually, their numbers, accumulated weight, and accumulated sizes.

This patch changes this so that the fair_queue expresses everything,
including its limits as fair_queue_tickets. We then enhance the
fair_queue_ticket to implement the <, += and -= operators and all
the relationship between them can then be expressed in a concise
and elegant manner.

One note for readers of this code: note how the cost is computed based
on the _config, while the decision to allow a request to pass through or
not is based on the current capacity. Those quantities are the same now,
but they are not guaranteed to be. I plan to allow the fair_queue to
grow and shrink its capacity.

Signed-off-by: Glauber Costa <glauber@scylladb.com>