fair_queue: return a pointer instead of a shared pointer for priority class
There is no need to return a shared pointer for the priority_class. We
can use a common technique used in other parts of the seastar code,
where we add all possible objects to a static container and then use a
stack to pile up the available entries.

To do that we do need to define a maximum amount of entries, but we have
a pretty good idea of what that would be: for instance, if the Fair
Queue is used to implement the I/O Queue, bursting through that max
would mean we need close to a 146-shard box sharing the same I/O queue
with 7 classes to reach it. If we attach one I/O Queue per shard that is
even less of a concern

The main disadvantage of this is that whenever we normalize the cost, we
would to iterate over a much larger array. Despite this being a rare
operation, in order to keep this cost in check we use an auxiliary set
with just the registered classes for this purpose.

Signed-off-by: Glauber Costa <glauber@scylladb.com>