fair_queue: allow registration and deregistration to happen across shards.
We are now dealing with naked pointers that are pre-allocated instead of
shared pointers. It should then be safe to register and de-register a
priority class from a foreign shard as long as we make sure that there
is no concurrent access to the data structures we use.

Register and unregister are rare operations. So in this patch we're not
being too smart, and just taking a lock. This could have been a mutex
but first, it doesn't make too much of a difference to us: if one of the
reactor sleeps with the mutex held it will block anyway. Second, we will
need a spinlock in dispatch for making the fair_queue multishard soon
anyway.

Signed-off-by: Glauber Costa <glauber@scylladb.com>