fair_queue: extract common elements from the fair_queue
In the next patch we will introduce a different version of the
fair_queue that is meant to work across shards. We want to do most of
the atomic work in a separate class to avoid forcing the fair_queue to
issue barriers but at the same time would like to share as much code as
possible.

This patch is a preparation patch whereas the elements that are planned
to be common are extracted in a class that both leaf classes will
inherit from.

Signed-off-by: Glauber Costa <glauber@scylladb.com>