multishard_fair_queue: an accounting fair_queue that can be used from many shards.
This patch introduces the multishard_fair_queue. It is an
accounting-only version of the fair queue that can be used to allow many
shard-local fair queues to consume resources from a shared pool.

To avoid bottlenecks, the multishard_fair_queue is designed to be
executed from any shard. It will try_lock on a lock and bail if another
shard is already executing. This is not overcommit friendly. However it
it is the fastest way to handle this in pinned scenarios, where seastar
is usually ran, so my main proposal for that is to simply do static
partitioning of resources (same as today) in overcommit scenarios.

The priority_class is augmented to contain two atomic quantities: an
ask, representing how much resources a fair_queue wants to consume, and
a grant, representing how much resources a fair_queue is allowed to
consume. Before computing a grant, all changes to ask must be visible.
We therefore use release / acquire semantics on ask and grant.

Ideally, we would be able to nest multishard_fair_queue. However my
attempts to nest it led to a lot of complication arising from the fact
that a multishard_fair_queue doesn't yet know what is ask is until it
tries to run over its fair_queues. Having the fair_queues update all
asks recursively at queue() is just not performant. It should be
possible to solve it by exchanging asks on a time-based fashion instead
at dispatch-time, with growing times as we go deeper in the hierarchy.
However to keep things moving I am just disallowing nesting for now.

One important difference between the fair_queue and the
multishard_fair_queue is that the multishard_fair_queue does not keep
its waiters in a priority queue. This would be impractical since the
whole point behind it is to batch operations. We therefore sort an array
of registered classes upon dispatch. The assumptions behind this are:
 - there will be a small, fixed number of elements in this array
 - all elements of the array will likely have activity at every dispatch

Those assumptions should hold in the common case in which the multishard
fair_queue is used to implement a unified I/O queue per all shards:
there will be one priority_class per shard at most (or even fewer if we
restrict this to a socket), and all shards should always want to
dispatch something if the system is not idle.

Signed-off-by: Glauber Costa <glauber@scylladb.com>